cover-display/requirements.txt
spotipy
python-dotenv
pillow
requests
